BOOKMARK commit
There are multiple issues with these schemas. We don't want to lose track of any of the desirable features as the issues are resolved. This commit serves as an easily discovered flag: at this point, no changes have yet been made, and there are no pending pull requests. DIFF THIS COMMIT AGAINST FUTURE STATES OF THIS REPO TO LOOK FOR ANYTHING THAT FELL THROUGH THE CRACKS.